Invoke address resolution upon subsequent traffic to Failed neighbor
Removes the period of time in which subseqeuent traffic to a Failed neighbor
immediately fails with ErrNoLinkAddress. A Failed neighbor is one in which
address resolution fails; or in other words, the neighbor's IP address cannot
be translated to a MAC address.

This means removing the Failed state for both neighborCache and linkAddrCache.
Previously, both caches would transition entries to Failed after address
resolution fails. In this state, any subsequent traffic requested within an
unreachable time would immediately fail with ErrNoLinkAddress. This does not
follow RFC 4861 section 7.3.3:

  If address resolution fails, the entry SHOULD be deleted, so that subsequent
  traffic to that neighbor invokes the next-hop determination procedure again.
  Invoking next-hop determination at this point ensures that alternate default
  routers are tried.

Fixes #4796

PiperOrigin-RevId: 341996543